+++ Modules/Intent.Modules.SqlDatabaseProject/release-notes.md
@@ -1,1 +1,3 @@
 ### Version 1.0.0
+
+- New Feature: Generate SQL Database Project containing SQL files from Domain Entities to product DACPAC files for DB deployment.